Safari-611.1.10/LayoutTests/imported/w3c/web-platform-tests/workers/SharedWorker_dataUrl-expected.txt

CONSOLE MESSAGE: ReferenceError: Can't find variable: SharedWorker
CONSOLE MESSAGE: ReferenceError: Can't find variable: SharedWorker


Harness Error (TIMEOUT), message = null

TIMEOUT Data URL not shared by cross-origin SharedWorkers Test timed out
NOTRUN Data URLs shared by same-origin SharedWorkers
